feat(trustscan): implement TrustScan service with domain, header, and SSL checks, including utility functions and response handling
